make all timesteps ints redefine the scope of simulation (remove functionality) improve runners

+++ src/general_support/Integrator.java
@@ -3,10 +3,10 @@ package general_support;
 public interface Integrator {
 
     /**
-     *
-     * @param position Vector indicating current position
-     * @param acceleration Vector indicating current acceleration
+     *  @param position Vector indicating current position
      * @param velocity Vector indicating current velocity
+     * @param acceleration Vector indicating current acceleration
+     * @param timeStep
      */
-    void integrate(Vector position, Vector velocity, Vector acceleration, double timeStep);
+    void integrate(Vector position, Vector velocity, Vector acceleration, int timeStep);
 }
